Refactor caching interface and helpers
